+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/add_ch_privilegies_original.sql
@@ -1,0 +1,4 @@
+CREATE USER timon
+	DEFAULT ROLE admin;
+
+CREATE ROLE admin;