+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/ch_ignore_privillegies_new.sql
@@ -1,0 +1,4 @@
+CREATE USER khazieva_gr
+	DEFAULT ROLE admin;
+
+GRANT SELECT(c3) ON default.table1 TO khazieva_gr WITH GRANT OPTION;