+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/ch_ignore_privillegies_original.sql
@@ -1,0 +1,6 @@
+CREATE USER khazieva_gr
+	DEFAULT ROLE admin;
+
+GRANT INSERT ON default.table1 TO khazieva_gr;
+GRANT SELECT(c1) ON default.table1 TO khazieva_gr WITH GRANT OPTION;
+GRANT SELECT(c2) ON default.table1 TO khazieva_gr WITH GRANT OPTION;