+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/revoke_ch_privilegies_new.sql
@@ -1,0 +1,6 @@
+--revoke privillegies for USER
+CREATE USER john
+	DEFAULT ROLE admin;
+
+--revoke privillegies for ROLE
+CREATE ROLE admin;